+++ Presentation/OpenTgResearcherDesktop/Helpers/TgResourceExtensions.cs
@@ -35,14 +35,15 @@ public static class TgResourceExtensions
 	internal static string AskAddRecord() => "AskAddRecord".GetLocalizedResource();
 	internal static string AskCalcChatStatistics() => "AskCalcChatStatistics".GetLocalizedResource();
 	internal static string AskCalcContentStatistics() => "AskCalcContentStatistics".GetLocalizedResource();
+	internal static string AskClearParticipants() => "AskClearParticipants".GetLocalizedResource();
 	internal static string AskClientConnect() => "AskClientConnect".GetLocalizedResource();
 	internal static string AskClientDisconnect() => "AskClientDisconnect".GetLocalizedResource();
 	internal static string AskDataClear() => "AskDataClear".GetLocalizedResource();
-	internal static string AskLoading() => "AskLoading".GetLocalizedResource();
 	internal static string AskDeleteFile() => "AskDeleteFile".GetLocalizedResource();
 	internal static string AskGetParticipants() => "AskGetParticipants".GetLocalizedResource();
 	internal static string AskLicenseCheck() => "AskLicenseCheck".GetLocalizedResource();
 	internal static string AskLicenseClear() => "AskLicenseClear".GetLocalizedResource();
+	internal static string AskLoading() => "AskLoading".GetLocalizedResource();
 	internal static string AskLicenseRequestCommunity() => "AskLicenseRequestCommunity".GetLocalizedResource();
 	internal static string AskRestartApp() => "AskRestartApp".GetLocalizedResource();
 	internal static string AskSettingsClear() => "AskSettingsClear".GetLocalizedResource();